+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Move tests to actual tests
+
 ## v0.1.0
 
 - Initial release